input extracted into its own package to make it more reusable
