+++ internal/tui/components/Input/config.go
@@ -1,0 +1,8 @@
+package input
+
+type InputConfig struct {
+	Prompt      string
+	Placeholder string
+	CharLimit   int
+	Width       int
+}
